# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: ManufactureOS ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workorders:
    get:
      operationId: listWorkOrders
      tags: [workorders]
      summary: List work orders
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: itemId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Work orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrdersListResponse"
    post:
      operationId: createWorkOrder
      tags: [workorders]
      summary: Create work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkOrderRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /workorders/{id}:
    get:
      operationId: getWorkOrder
      tags: [workorders]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Work order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
    put:
      operationId: updateWorkOrder
      tags: [workorders]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkOrderRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /workorders/{id}/complete:
    post:
      operationId: completeWorkOrder
      tags: [workorders]
      summary: Complete work order
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteWorkOrderRequest"
      responses:
        "200":
          description: Work order completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # SHIPMENTS
...
  /mrp/runs:
    get:
      operationId: listMrpRuns
      tags: [mrp]
      summary: List MRP runs
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: MRP runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MrpRunsListResponse"
    post:
      operationId: runMrp
      tags: [mrp]
      summary: Trigger MRP run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunMrpRequest"
      responses:
        "201":
          description: MRP run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MrpRun"
  /mrp/runs/{id}:
    get:
      operationId: getMrpRun
      tags: [mrp]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: MRP run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MrpRun"
  /mrp/recommendations:
    get:
      operationId: listMrpRecommendations
      tags: [mrp]
      summary: List MRP recommendations
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: runId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MrpRecommendationsListResponse"

  # PLANNING / WHAT-IF
  /planning/scenarios:
    get:
      operationId: listPlanningScenarios
      tags: [planning]
      summary: List planning scenarios
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Scenarios
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanningScenariosListResponse"
    post:
      operationId: createPlanningScenario
      tags: [planning]
      summary: Create planning scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanningScenarioRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanningScenario"
  /planning/scenarios/{id}:
    get:
      operationId: getPlanningScenario
      tags: [planning]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanningScenario"

  # DASHBOARD
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 25
...
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages

    # USER / AUTH
...
    WorkOrderOperation:
      type: object
      properties:
        id:
          type: string
        sequence:
          type: integer
        name:
          type: string
        workcenterId:
          type: string
        workcenterName:
          type: string
        status:
          type: string
          enum: [pending, in_progress, complete, skipped]
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        actualStart:
          type: string
          format: date-time
        actualEnd:
          type: string
          format: date-time
        setupTime:
          type: number
        runTime:
          type: number
        laborHours:
          type: number
      required:
        - id
        - sequence
        - name
        - status
    WorkOrder:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        itemId:
          type: string
        itemNumber:
          type: string
        itemName:
          type: string
        bomId:
          type: string
        routingId:
          type: string
        salesOrderId:
          type: string
        salesOrderNumber:
          type: string
        status:
          type: string
          enum: [draft, released, in_progress, complete, cancelled, on_hold]
        type:
          type: string
          enum: [standard, rework, repair]
        quantityOrdered:
          type: number
        quantityCompleted:
          type: number
        quantityScrapped:
          type: number
        uom:
          type: string
        scheduledStart:
          type: string
          format: date
        scheduledEnd:
          type: string
          format: date
        actualStart:
          type: string
          format: date
        actualEnd:
          type: string
          format: date
        operations:
          type: array
          items:
            $ref: "#/components/schemas/WorkOrderOperation"
        warehouseId:
          type: string
        priority:
          type: string
          enum: [low, normal, high, urgent]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - number
        - itemId
        - status
        - quantityOrdered
    CreateWorkOrderRequest:
      type: object
      properties:
        itemId:
          type: string
        bomId:
          type: string
        routingId:
          type: string
        salesOrderId:
          type: string
        type:
          type: string
          enum: [standard, rework, repair]
        quantityOrdered:
          type: number
        scheduledStart:
          type: string
          format: date
        scheduledEnd:
          type: string
          format: date
        warehouseId:
          type: string
        priority:
          type: string
          enum: [low, normal, high, urgent]
        notes:
          type: string
      required:
        - itemId
        - quantityOrdered
    UpdateWorkOrderRequest:
      type: object
      properties:
        status:
          type: string
          enum: [released, on_hold, cancelled]
        scheduledStart:
          type: string
          format: date
        scheduledEnd:
          type: string
          format: date
        priority:
          type: string
          enum: [low, normal, high, urgent]
        notes:
          type: string
    CompleteWorkOrderRequest:
      type: object
      properties:
        quantityCompleted:
          type: number
        quantityScrapped:
          type: number
        actualEnd:
          type: string
          format: date
        notes:
          type: string
      required:
        - quantityCompleted
    WorkOrdersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WorkOrder"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta

    # SHIPMENT
...
    MrpRun:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [queued, running, complete, failed]
        type:
          type: string
          enum: [full_regen, net_change]
        planningHorizon:
          type: integer
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        summaryStats:
          type: object
          properties:
            plannedPOs:
              type: integer
            plannedWOs:
              type: integer
            shortages:
              type: integer
            rescheduleMessages:
              type: integer
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - type
    RunMrpRequest:
      type: object
      properties:
        type:
          type: string
          enum: [full_regen, net_change]
        planningHorizon:
          type: integer
      required:
        - type
    MrpRecommendation:
      type: object
      properties:
        id:
          type: string
        runId:
          type: string
        type:
          type: string
          enum: [planned_po, planned_wo, reschedule, shortage, cancel]
        itemId:
          type: string
        itemNumber:
          type: string
        itemName:
          type: string
        quantity:
          type: number
        neededDate:
          type: string
          format: date
        vendorId:
          type: string
        priority:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [open, released, accepted, rejected]
      required:
        - id
        - runId
        - type
        - itemId
        - quantity
    MrpRunsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MrpRun"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    MrpRecommendationsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MrpRecommendation"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta

    # PLANNING SCENARIOS
    PlanningScenario:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, active, published, archived]
        baselineRunId:
          type: string
        modifications:
          type: object
        comparisonResult:
          type: object
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
    CreatePlanningScenarioRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        baselineRunId:
          type: string
        modifications:
          type: object
      required:
        - name
    PlanningScenariosListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PlanningScenario"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta

    # DASHBOARD
...